# Add caching for fetching object by keys