The PR adds caching for fetching objects

It allows to speed up expressions where more than one method should be called 
Before
```
Benchmark                                       Mode  Cnt     Score    Error   Units
DatafakerBenchmark.address                     thrpt    5   109.668 ±  1.462  ops/ms
DatafakerBenchmark.oneMethodCallExpression     thrpt    5  1087.463 ± 24.951  ops/ms
DatafakerBenchmark.threeMethodsCallExpression  thrpt    5   281.641 ±  7.682  ops/ms
```

After
```
Benchmark                                       Mode  Cnt     Score    Error   Units
DatafakerBenchmark.address                     thrpt    5   119.390 ±  1.479  ops/ms
DatafakerBenchmark.oneMethodCallExpression     thrpt    5  1088.493 ± 19.085  ops/ms
DatafakerBenchmark.threeMethodsCallExpression  thrpt    5   328.149 ±  7.412  ops/ms
```
For one method it brings nothing however for cases where more than one method should be called (under fullAddress there are several methods should be called) it brings about 10% speed up